+++ src/styles/Home.module.css
@@ -106,6 +106,47 @@
 .logo {
   position: relative;
 }
+
+.upcomingEventHeader {
+  display: flex;
+  align-items: center;
+  margin-left: 2rem;
+  gap: 0.25rem;
+}
+
+.upcomingEventHeader::before {
+  content: '';
+  height: 24px;
+  width: 24px;
+  margin-right: 1rem;
+  background-image: url("../../assets/design-vectors/arrow.svg");
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+
+.upcomingEventHeader::after {
+  content: '';
+  height: 26px;
+  width: 26px;
+  background-image: url("../../assets/design-vectors/cursor.svg");
+  background-repeat: no-repeat;
+}
+
+.eventContainer {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1.5rem;
+  align-items: center;
+  justify-content: center;
+  padding: 2rem;
+  position: relative;
+  background-image: url("../../assets/design-vectors/grid.svg");
+}
+
+.officeHours {
+  text-align: center;
+}
+
 /* Enable hover only on non-touch devices */
 @media (hover: hover) and (pointer: fine) {
   .card:hover {
